moving mpp loss into wrapper
