+++ vit_pytorch/__init__.py
@@ -1,3 +1,2 @@
 from vit_pytorch.vit_pytorch import ViT
-from vit_pytorch.mpp_loss import MPPLoss
 from vit_pytorch.mpp_pytorch import MPP
